checkboxes fully working, degreereport.py merged with server.py
