+++ src/Components/Card/DegreeReport.js
@@ -12,17 +12,15 @@ import Image from 'react-bootstrap/Image'
 import Logo from '../static/images/logo.jpg'
 
 global.email = String(window.location).split("?")[1];
-global.email = String(global.email).split("=")[1]
+global.email = String(global.email).split("=")[1];
 
 export const Report = () => {
-    
     const [isLoading, setLoading] = useState(true);
     const [success, setSuccess] = useState();
-    const[show, setShow] = useState(false);
 
-    const [checked, setChecked] = useState([]);
+    const [checked, setChecked] = useState();
     const [unchecked, setUnchecked] = useState([]);
-    const [selected, setSelected] = useState([]);
+    const [selected, setSelected] = useState();
     const [unselected, setUnselected] = useState([]);
 
     useEffect(() => {
@@ -32,7 +30,7 @@ export const Report = () => {
         axios.get("/api/degreereport?email=" + global.email).then(response => {
             setSuccess(response.data);
             setChecked(response.data[2]["checked"]);
-            setSelected(response.data[2]["selected"])
+            setSelected(response.data[2]["selected"]);
             setLoading(false);
         });
     }, []);
@@ -54,53 +52,29 @@ export const Report = () => {
         })
     }
 
-    if (isLoading) {
-        return <div> Loading... </div>
-    }
-
     /**
-     * Checks if a course (under the checkboxes) is already in the database 
-     * @param {*} courseCode the code of the course we are checking if complete
-     * @param {*} reqCat    the requirement category under which the course falls
+     * Checks if a course (under the checkboxes) is already in the database
+     * @param {*} courseCode  the code of the course we are checking if complete
+     * @param {*} reqCat the requirement category under which the course falls
      * @returns true if the course has been checked, false otherwise
      */
     // THIS WORKS AND IS TESTED
     function isChecked(courseCode, reqCat) {
         for(const course of checked) {
-            if(course["course_code"] === courseCode && course["req_category"] === reqCat) {
+            if(course["course_code"] == courseCode && course["req_category"] == reqCat) {
                 return true
             }
         }
+
         return false
     }
-
-    // const handleChecked = (event, courseCode, courseName, reqCat, reqYear) => {
-    //     if(event.target.checked && (checked.some(e => e.course_code == courseCode) || selected.some(e => e.course_code == courseCode))) {
-    //         alert(courseCode + " " + courseName + " is already marked as completed.")
-    //     }
-    //     else if(event.target.checked && checked.some(e => e.course_code == courseCode) === false && selected.some(e => e.courseCode == courseCode) === false) {
-    //         setUnchecked(unchecked.filter((course) => course.course_code !== courseCode || course.req_category !== reqCat))
-    //         setUnselected(unselected.filter((course) => course.course_code !== courseCode || course.req_category !== reqCat))
-    //         setChecked(checked.concat({course_code: courseCode, course_name: courseName, req_category: reqCat, req_yr: reqYear}))
-    //     }
-    //     else if(event.target.unchecked) {
-    //         setChecked(checked.filter((course) => course.course_code !== courseCode || course.req_category !== reqCat))
-    //         setUnchecked(unchecked.concat({course_code: courseCode, course_name: courseName, req_category: reqCat, req_yr: reqYear}))
-    //     }
-
-    //     console.log("Checked: ")
-    //     console.log(checked)
-
-    //     console.log("Uhchecked: ")
-    //     console.log(unchecked)
-    // }
 
     /**
      * Grabs all courses under the given requirement category that have been completed
      * @param {*} reqCat the requirement category that we want to check if any completed courses fall under
-     * @returns a list of courses (in course code then course name format) that have been completed for that requirement category
+     * @returns a list of courses (format: courseCode + courseName) that have been completed for that requirement category
      */
-    //THIS WORKS AND HAS BEEN TESTED!!!
+    //THIS WORKS AND HAS BEEN TESTED!!
     function isSelected(reqCat) {
         var selectedCourses = []
 
@@ -111,6 +85,45 @@ export const Report = () => {
         }
 
         return selectedCourses
+    }
+
+    /**
+     * Event handler for when a box is checked or unchecked,
+     * Adds checked courses to a list of completed courses and removes if previously marked as unchecked
+     * Adds unchecked courses to list of previously completed courses and removes from current completed list
+     * @param {} event the event that happened (i.e., box was checked or unchecked)
+     * @param {*} courseCode the course code
+     * @param {*} courseName the name of the course
+     * @param {*} reqCat the requirement category that the course falls under
+     * @param {*} reqYear the requirement year of the student
+     */
+    //WORKS, NEEDS TO BE TESTED MORE
+    const handleChecked = (event, courseCode, courseName, reqCat, reqYear) => {
+        //Checks if course has already been marked as completed in some other category and alerts the user if so
+        if(event.target.checked && (checked.some(e => e.course_code == courseCode && e.req_category != reqCat) || selected.some(e => e.course_code == courseCode))) {
+            alert(courseCode + " " + courseName + " has already been marked as completed.")
+        }
+        //handles when course was marked as completed
+        else if(event.target.checked) {
+            setUnchecked(unchecked.filter((course) => course.course_code !== courseCode || course.req_category !== reqCat))
+            setUnselected(unselected.filter((course) => course.course_code !== courseCode || course.req_category !== reqCat))
+            setChecked(checked.concat({course_code: courseCode, course_name: courseName, req_category: reqCat, req_yr: reqYear}))
+        }
+        //handles when course was previously marked as complete and is now incomplete
+        else if(!event.target.checked) {
+            setChecked(checked.filter((course) => course.course_code !== courseCode || course.req_category !== reqCat))
+            setUnchecked(unchecked.concat({course_code: courseCode, course_name: courseName, req_category: reqCat, req_yr: reqYear}))
+        }
+
+        console.log("Checked: ")
+        console.log(checked)
+
+        console.log("Unchecked: ")
+        console.log(unchecked)
+    }
+
+    if(isLoading) {
+        return <div> Loading... </div>
     }
 
     return (
@@ -132,76 +145,65 @@ export const Report = () => {
                 <meta charSet="UTF-8"></meta>
             </div>
 
-            <h1>{ success[0]["degree_name"] }</h1>
+            <h1> { success[0]["degree_name"] } </h1>
 
             <div>
-                <h3> {"Total Hours: "} { success[0]["degree_hours"] }</h3>
+                <h3> { "Total Hours: "} { success[0]["degree_hours"] } </h3>  
+                <Form id="degree_report">      
+                <Button variant="primary" value="Submit" onClick={submitListener}>
+                            Save Changes
+                </Button>
 
-                <Form id="degree_report">
-                    <Button variant="primary" type="submit" value="Submit" onClick={submitListener}>Save Changes</Button>
-
-                    { success[0]["req_details"].map((req) => {
-                        return <div>
-                            {req["req_courses"].length > 0 ?
-                                <div className="col-md-12">
-                                    <Form.Group>
-                                        <h2> {req["req_category"]} </h2>
-                                        <h4> {"Required Hours: "} {req["required_hrs"]} </h4>
-                                        <p> { req["req_details"] }</p>
-
+                { success[0]["req_details"].map((req) => {
+                    return <div>
+                        {req["req_courses"].length > 0 ? 
+                            <div className="col-md-12">
+                                <Form.Group>
+                                    <h2> {req["req_category"]} </h2> 
+                                    <h4> {"Required Hours: "} {req["required_hrs"]} </h4>
+                                    <p> {req["req_details"]} </p>
                                         { req["req_courses"].map((course) => {
                                             return <FormControlLabel control = {
                                                 <Checkbox
                                                     defaultChecked = {isChecked(course["course_code"], req["req_category"])}
-                                                    name = {course["course_code"] + " " + course["course_name"]}
-                                                    // onChange = {(event) => handleChecked(event, course["course_code"], course["course_name"], req["req_category"], success[0]["req_yr"])}
+                                                    name = {req["req_category"] + " " + course["course_name"]}
+                                                    onChange={(event) => handleChecked(event, course["course_code"], course["course_name"], req["req_category"], success[0]["req_yr"])}
                                                 />
                                             }
                                             label = {course["course_code"] + " " + course["course_name"]}/>
                                         })}
-                                    </Form.Group>
-                                </div>
+                                </Form.Group>
+                            </div>
                             : <div className="col-md-12">
                                 <Form.Group>
-                                    <h2> {req["req_category"]} </h2>
+                                    <h2> { req["req_category"] }</h2>
                                     <h4> {"Required Hours: "} { req["required_hrs"]} </h4>
                                     <p> { req["req_details"] }</p>
                                     <div>
-                                        <Autocomplete 
+                                        <Autocomplete
                                             defaultValue = {isSelected(req["req_category"])}
                                             id="elective-courses"
                                             multiple
                                             options={success[1].map((course) => course["course_code"] + " " + course["course_name"])}
-                                            style={{ width: 600}}
+                                            style={{width: 600}}
                                             renderInput = {(params) =>
                                                 <TextField {...params}
-                                                    label="Select course"
+                                                    label="Enter course"
                                                     variant="outlined"
                                                 />
                                             }
                                         />
                                     </div>
                                 </Form.Group>
-                            
                             </div>
                         }
-
-                        </div>
-                    })}
-
-
-
-
-
-
-                </Form>
-
+                    </div>
+                })}
+            </Form>
             </div>
-
-
         </div>
-
     );
 }
 
 export default Report
+
